// src/models/interfaces/user.ts
export interface User {
	userID: string;
	role: string;
	personContact: string;
	mobileNumber: string;
	email: string;
	businessName: string;
	streetAddress1: string;
	streetAddress2: string;
	city: string;
	state: string;
	country: string;
	pincode: string;
	password: string;
	planID: string;
	created: number;
}